assign user to a task

+++ src/tasks/tasks.service.ts
@@ -7,7 +7,6 @@ import {Task} from "./task.entity";
 import {TaskStatus} from "./task-status-enum";
 import {GetTaskFilterDto} from "./dto/GetTaskFilterDto";
 import {User} from "../auth/user.entity";
-
 @Injectable()
 export class TasksService {
     constructor(
@@ -15,34 +14,37 @@ export class TasksService {
         private taskRepository: TaskRepository) {
     }
 
-    async getTasks(filterDTO: GetTaskFilterDto) {
-        return this.taskRepository.getTasks(filterDTO);
+    async getTasks(filterDTO: GetTaskFilterDto, user: User) {
+        return this.taskRepository.getTasks(filterDTO, user);
     }
 
-    async getTaskById(id: number) : Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User)
+        : Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: {id, userId: user.id}});
 
         if(!found) {
             throw new NotFoundException(`Task not found`);
         }
 
-        return  found
+        return  found;
     }
 
     async createTask(createTaskDto : CreateTaskDto, user: User): Promise<Task> {
         return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(taskId: number) : Promise<void> {
-        const result = await this.taskRepository.delete(taskId)
+    async deleteTaskById(taskId: number,  user: User) : Promise<void> {
+        const result = await this.taskRepository.delete({id: taskId, userId: user.id});
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task:Task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus,user: User): Promise<Task> {
+        const task:Task = await this.getTaskById(id, user);
 
         task.status = status;
